+++ R/reexports.R
@@ -21,3 +21,7 @@ hardhat::extract_mold
 #' @importFrom hardhat extract_preprocessor
 #' @export
 hardhat::extract_preprocessor
+#'
+#' @importFrom hardhat extract_parameter_set_dials
+#' @export
+hardhat::extract_parameter_set_dials